% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batteryCheck.R
\name{batteryCheck}
\alias{batteryCheck}
\title{Battery check}
\usage{
batteryCheck(raw_data)
}
\arguments{
\item{raw_data}{The raw data obtained from \code{readTTdata}}
}
\description{
Allows the user to check the battery level of the devices
  and the last time they sent a string.
}
\examples{
\dontrun{
# example battery check
data(raw4d_ex_data)
batteryCheck(raw4d_ex_data)
}



}
