% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSTWC.R
\name{calculateSTWC}
\alias{calculateSTWC}
\title{Calculate Stem Water Content}
\usage{
calculateSTWC(Ecf_Hz, Tref0, species = "Fagus")
}
\arguments{
\item{Ecf_Hz}{The frequency measured by the probe for estimating stem water
content.}

\item{Tref0}{The temperature (C) measured by the reference probe inside the
trunk.}

\item{species}{"Fagus" or "Pinus". Equations for calculating the Stem Water
Content are species specific, so you must specify the species to use the
right equation. Equations are available only for Fagus sylvatica and Pinus
sylvestris at the moment. If an invalid string is entered the function
raises an error.}
}
\value{
A numeric vector with the % volumetric water content, estimated
  according to the equations in the TT+ manual.
}
\description{
This function calculates the Stem Water Content from the TT data
  Nowdays formulas are available only for Fagus sylvatica and Pinus sylvestris.
}
\examples{
\dontrun{
#Primarily an internal function of the package
# Usually the user doesn't need to call it
  calculateSTWC(
  Ecf_Hz = 11383,
  Tref0 = 21.3,
  species = "Fagus" )
}


}
