% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTheat10.R
\name{calculateTheat10}
\alias{calculateTheat10}
\title{Calculate Theat 10}
\usage{
calculateTheat10(data)
}
\arguments{
\item{data}{Data of a single TT+ devices, arranged by \code{date_hour}.}
}
\value{
The same data.frame as data, plus a column called \code{Theat10} 
  containing the estimated value.
}
\description{
Calculates the temperature of the trunk 10 minutes after the 
  heating start, as it was if no heating from the probe occurred. The value is 
  computed by linear interpolation with Theat0 of the following hour. The procedure 
  allows for correcting for changes in trunk temperature, due to daily cycles.
}
\examples{
# Internal function of the package, the user does not need to call it

}
