% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string45Handling.R
\name{string45Handling}
\alias{string45Handling}
\title{Function for handling the string 45 data}
\usage{
string45Handling(
  TTdata,
  lower.TTree = 0,
  higher.TTree = 40,
  lower.TAir = -15,
  higher.TAir = 50,
  lower.RH = 35,
  higher.RH = 100,
  lower.sap.flow = 0,
  higher.sap.flow = 1080,
  remove.raw = "yes",
  rectangular = "no"
)
}
\arguments{
\item{TTdata}{Raw ata from the Tree Talker devices. The data returned by the 
readTTdata function.}

\item{lower.TTree}{Lower threshold of temperature inside the trunk. 
Values lower than this will be regarded as non valid and replaced with NA.
Default value = \code{0}.}

\item{higher.TTree}{Higher threshold of temperature inside the trunk. 
Values higher than this will be regarded as non valid and replaced with NA.
Default value = \code{40}.}

\item{lower.TAir}{Lower threshold of air temperature regarded as valid. 
Values lower than this will be regarded as non valid and replaced with NA.
Default value = \code{-15}.}

\item{higher.TAir}{Higher threshold of air temperature regarded as valid. 
Values higher than this will be regarded as non valid and replaced with NA.
Default value = \code{50}.}

\item{lower.RH}{Lower threshold of Relative Humidity (%) regarded as valid. 
Values lower than this will be regarded as non valid and replaced with NA.
Default value = \code{35}.}

\item{higher.RH}{Higher threshold of Relative Humidity (%) regarded as valid. 
Values higher than this will be regarded as non valid and replaced with NA.
Default value = \code{100}}

\item{lower.sap.flow}{Lower threshold of sap flow. Values lower than this will 
be replaced with NA. Default value = \code{0}.}

\item{higher.sap.flow}{Higher threshold of sap flow. Values higher than this will be 
considered invalid and replaced with NA. Default value = \code{1080}.}

\item{rectangular}{If to retain dates and hours of missing strings. If \code{"no"}
they will be discarded, getting lighter data. In any other case the full grid of
ids x times will be kept.}

\item{tz}{Timezone of the site analyzed, to properly convert the timestamp in
the local time value. Defaults to \code{Sys.timezone}.}

\item{remove_raw}{If to remove columns of raw data or retain them. Default value:
 \code{"yes"}
to remove them, in any other case they will be retained.}
}
\value{
A data frame with TT+ 4D data: air temperature, relative humidity, vpd, 
  temperatures of the sap flow probes, K index, sap flux estimated with the 
  Do et al. (2011) equation, for two and one probe, sap flux estimated with the 
  equation in Asgharinia et al. (2022), growth raw data (DN), stem Water content
  sensor raw values and values calculated using the species specific calibration 
  equations.
}
\description{
Function for handling the string 45 of the old Tree Talkers. 
The function is very similar to the \code{clean4DData} function in its inputs and outputs.
}
\section{Functions}{
\itemize{
\item \code{string45Handling()}: 

}}
\references{
Do, F.C.; Isarangkool Na Ayutthaya, S.; Rocheteau, A. Transient thermal dissipation method for xylem sap flow measurement: Implementation with a single probe. Tree Physiol. 2011, 31, 369– 380, doi:10.1093/treephys/tpr020.
  Asgharinia, S., Leberecht, M., Belelli Marchesini, L., Friess, N., Gianelle, D., Nauss, T., ... & Valentini, R. (2022). Towards Continuous Stem Water Content and Sap Flux Density Monitoring: IoT-Based Solution for Detecting Changes in Stem Water Dynamics. Forests, 13(7), 1040.
}
