% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean4DData.R
\name{clean4DData}
\alias{clean4DData}
\title{Clean 4D data}
\usage{
clean4DData(
  data,
  lower.TTree = 0,
  higher.TTree = 40,
  lower.TAir = -15,
  higher.TAir = 50,
  lower.RH = 35,
  higher.RH = 100,
  lower.sap.flow = 0,
  higher.sap.flow = 1080,
  lower.VWC = 0,
  higher.VWC = 100,
  species = "FagusS"
)
}
\description{
This is the main function in this package. Its purpose is to
  filter and convert the 4D data from the TT+ raw data downloaded from the server
  The data are converted from DNs to physical quantities using the equations in the
  TT+ manual. The data are also cleaned, removing values that are regarded
  not to be plausible (e.g. temperatures that are too high, too low etc.) by
  calling the helper functions.
}
