% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudData4B.R
\name{cloudData4B}
\alias{cloudData4B}
\title{TT cloud 4B string processing}
\usage{
cloudData4B(raw_data)
}
\arguments{
\item{raw_data}{The raw Tree Talkers data, as those obtained by the readTTData
function.}
}
\value{
A data frame with the 4B data properly organized.
}
\description{
this function isolates the 4B string, renames its columns,
  retrieves the id and converts the timestamp to date-time.
}
\examples{
\dontrun{
data(raw45_ex_data)
cloudData4B(raw45_ex_data)
}

}
