% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertStringsEnsemble.R
\name{convertStringsEnsemble}
\alias{convertStringsEnsemble}
\title{Function for converting the 4D, 45 and 59 strings all together}
\usage{
convertStringsEnsemble(
  data,
  lower.TTree = 0,
  higher.TTree = 40,
  lower.TAir = -15,
  higher.TAir = 50,
  lower.RH = 35,
  higher.RH = 100,
  lower.sap.flow = 0,
  higher.sap.flow = 300,
  lower.VWC = 0,
  higher.VWC = 100,
  species = "Fagus",
  tz = Sys.timezone(),
  startdate,
  enddate
)
}
\arguments{
\item{data}{Tree talker data, read with the \code{readTTData} function of the package.}

\item{lower.TTree}{Lower threshold of temperature inside the trunk. 
Values lower than this will be regarded as non valid and replaced with NA.
Default value = \code{0}.}

\item{higher.TTree}{Higher threshold of temperature inside the trunk. 
Values higher than this will be regarded as non valid and replaced with NA.
Default value = \code{40}.}

\item{lower.TAir}{Lower threshold of air temperature regarded as valid. 
Values lower than this will be regarded as non valid and replaced with NA.
Default value = \code{-15}.}

\item{higher.TAir}{Higher threshold of air temperature regarded as valid. 
Values higher than this will be regarded as non valid and replaced with NA.
Default value = \code{50}.}

\item{lower.RH}{Lower threshold of Relative Humidity (%) regarded as valid. 
Values lower than this will be regarded as non valid and replaced with NA.
Default value = \code{35}.}

\item{higher.RH}{Higher threshold of Relative Humidity (%) regarded as valid. 
Values higher than this will be regarded as non valid and replaced with NA.
Default value = \code{100}}

\item{lower.sap.flow}{Lower threshold of sap flow. Values lower than this will 
be replaced with NA. Default value = \code{0}.}

\item{higher.sap.flow}{Higher threshold of sap flow. Values higher than this will be 
considered invalid and replaced with NA. Default value = \code{1080}.}

\item{lower.VWC}{Lower threshold for volumetric water content. Values lower than
this will be considered as non valid and replaced with NA. Default value = 
\code{0}.}

\item{higher.VWC}{Higher threshold for volumetric water content. Values higher 
than this will be considered as non valid and replaced with NA. 
Default value = \code{100}.}

\item{species}{Species the TT+ are installed on. Needed to retrieve the 
species specific coefficient for Stem Water Content estimation. 
Possible values are \code{"Fagus"} and \code{"Oak"}, with \code{"Fagus"}
as default value. As more calibrations become available they will be added.}

\item{tz}{Timezone of the site analyzed, to properly convert the timestamp in
the local time value. Defaults to \code{Sys.timezone}.}
}
\value{
A data frame with TT+ 4D data: air temperature, relative humidity, vpd, 
  temperatures of the sap flow probes, K index, sap flux estimated with the 
  Do et al. (2011) equation, for two and one probe, sap flux estimated with the 
  equation in Asgharinia et al. (2022), growth raw data (DN), stem Water content
  sensor raw values and values calculated using the species specific calibration 
  equations.
}
\description{
Wrapper for the clean4DData, string45Handling and clean59Data
  functions, Its main purpose is to power the launchNewGUI function, and to
  ease the conversion of the TT data in clouds where different types of TT 
  devices are mixed.
}
\examples{
# example code
\dontrun{
data(raw45_ex_data)
data(raw4d_ex_data)
data(raw59_ex_data)
convertStringsEnsemble(rbind(rbind(raw45_ex_data, raw4d_ex_data), raw59_ex_data))
}

}
\references{
Do, F.C.; Isarangkool Na Ayutthaya, S.; Rocheteau, A. Transient thermal dissipation method for xylem sap flow measurement: Implementation with a single probe. Tree Physiol. 2011, 31, 369 380, doi:10.1093/treephys/tpr020.

  Asgharinia, S., Leberecht, M., Belelli Marchesini, L., Friess, N., Gianelle, D., Nauss, T., ... & Valentini, R. (2022). Towards Continuous Stem Water Content and Sap Flux Density Monitoring: IoT-Based Solution for Detecting Changes in Stem Water Dynamics. Forests, 13(7), 1040.
}
