% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyNumberTTs.R
\name{dailyNumberTTs}
\alias{dailyNumberTTs}
\title{Tree talkers with valid data per day}
\usage{
dailyNumberTTs(
  qcdata,
  variable_of_interest = "do_sap_flow",
  min_valid_obs = 24
)
}
\arguments{
\item{qcdata}{Dataframe with quality check data, produced by 
\code{dailyQualityCheck}.}

\item{variable_of_interest}{String, the variable of interest. Must be one of
the colnames of qcdata. In the present version one of \code{"Tref0", "Theat0", "Tref1", "Theat1", "voltage", "Tair",
"RH", "do_sap_flow", "vpd", "sap_flow_one_probe", "asgharinia_sap_flow"}. 
Default value \code{"do_sap_flow"}.}

\item{min_valid_obs}{Numeric, minimum of valid observations per day required
for each device, to consider its data valid. Default value \code{24}.}
}
\value{
A dataframe with the number of TT with valid data.
}
\description{
Function that computes the number of TT with valid data per day,
  given the output of the \code{dailyQualityCheck} function output
}
