% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{detectOutliers}
\alias{detectOutliers}
\title{Flag ligth data outliers}
\usage{
detectOutliers(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A logical vector containing the flags.
}
\description{
This function flags light values as outliers according to the 
  "boxplot rule". In fact it flags as outliers all values that fall outside
  of the box and whisker of a box and whisker plot. The function computes the 
  0.75 and 0.25 quantiles of the data. A value is considered an outlier if it 
  falls outside the interval: \code{(Q25-1.5*IQR, Q75+1.5IQR)} and so is 
  flagged as outlier. This function takes a numeric vector as input and 
  returns a logical vector containing that flags (T means a value is an 
  outlier, F means it's not an outlier).
}
\examples{
\dontrun{

example_data <- rnorm(100, 2,3)
detectOutliers(example_data)

}

}
