% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviceFunctioningAssesment.R,
%   R/deviceFunctioningAssessment.R
\name{deviceFunctioningAssessment}
\alias{deviceFunctioningAssessment}
\title{Graphical report of TT+ technical funtioning}
\usage{
deviceFunctioningAssessment(
  TTdata4d,
  id,
  start_date = "2021-01-01",
  end_date = "2021-12-31",
  min_valid_obs = 24
)

deviceFunctioningAssessment(
  TTdata4d,
  id,
  start_date = "2021-01-01",
  end_date = "2021-12-31",
  min_valid_obs = 24
)
}
\arguments{
\item{TTdata4d}{Tree talker 4D data, as the dataframe produced by the
\code{clean4DData}) function.}

\item{id}{Device id as string. Must be one of the values stored in the 
4d data column id.}

\item{start_date}{Date of start of the time frame of interest.}

\item{end_date}{Date of end of the time frame of interest.}

\item{min_valid_obs}{Minimum valid observations of each variable to label the 
data of the corresponding date as ok (and display in green color)}
}
\value{
A ggplot object, the plot with the graphical assesment of the functioning
  of the device corresponding to that ID.
}
\description{
This function creates a graphical report of dayly probe state and
  device battery voltage for a specific Tree talker during a time specified by
  the user.

This function creates a graphical report of dayly probe state and
  device battery voltage for a specific Tree talker during a time specified by
  the user.
}
\examples{
\dontrun{
data(raw4d_ex_data)
clean_data <- clean4DData(raw4d_ex_data)
assessment_plot <- deviceFunctioningAssessment(
clean_data,
id = "11111111",
start_date = "2021-05-07",
end_date = "2021-05-20",
min_valid_obs = 23)
#the plot has can be printed
print(assessment_plot)
}

\dontrun{
data(raw4d_ex_data)
clean_data <- clean4DData(raw4d_ex_data)
assessment_plot <- deviceFunctioningAssessment(
clean_data,
id = "11111111",
start_date = "2021-05-07",
end_date = "2021-05-20",
min_valid_obs = 23)
#the plot has can be printed
print(assessment_plot)
}

}
