% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByHour.R
\name{filterByHour}
\alias{filterByHour}
\title{Filter by hour}
\usage{
filterByHour(data, hours)
}
\arguments{
\item{hours}{A character vectors of the hours to be included: e.g.:
c("2","3","4") or "12".}

\item{data4D}{The 4D TT+ data}
}
\description{
filter that allows the user to filter the TT data by hour.
}
