% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterIDs.R
\name{filterIDs}
\alias{filterIDs}
\title{Filter IDs}
\usage{
filterIDs(data4D, ids = "")
}
\arguments{
\item{data4D}{The data frame with 4D data}

\item{ids}{A vector with the ids of interest}
}
\description{
Function for filtering only certain ids from the dataset. Based 
  on the dplyr::filter().
}
