% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagOutliers.R
\name{flagOutliers}
\alias{flagOutliers}
\title{Flag outliers}
\usage{
flagOutliers(TTdata, variables, timeperiod = "month")
}
\arguments{
\item{TTdata}{Dataframe with TreeTalker data}

\item{variables}{Variables to which to apply the function.}

\item{timeperiod}{Flag (\code{"month" or "week" or "all"}). if TRUE data are grouped by week, 
else are grouped by month for applying the procedure.}
}
\description{
This function creates new columns in the TT+ data dataframe,
  that are outlier flags. It exploits the detectOutliers function, which 
  applies the boxplot rule. Data are grouped by hour and by week or month,
  and then the function in applied.
}
