% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatBatteryTable.R
\name{formatBatteryTable}
\alias{formatBatteryTable}
\title{Format battery table}
\usage{
formatBatteryTable(battery_table)
}
\arguments{
\item{battery_table}{The battery table (the output of fastBatteryCheck)}
}
\value{
The battery table properly formatted, with all columns converted
  to character.
}
\description{
Formats the battery table returned by \code{fastBatteryCheck}
  for properly displaying it in the GUI. The columns are converted to
  characters.
}
