% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthElaboration.R
\name{growthElaboration}
\alias{growthElaboration}
\title{Growth data elaboration}
\usage{
growthElaboration(data4d, nweeks)
}
\arguments{
\item{data4d}{The dataframe returned by \code{clean4DData}, with inside the
growth_DN column}

\item{nweeks}{The number of consecutive weeks to aggregate together. Must be
1, 2 or 3. The function might accept also higher values, but we don't
recommend to use them as the procedure has not been tested for those values.}
}
\description{
The raw DNs from the growth sensor are aggregated weekly,
  biweekly or by intervals of three weeks.
  Then the highest mode is taken. This value is than converted to measurment
  in cms. For each id the initial value is subtracted to all the values measured.
}
\examples{
\dontrun{
data(raw4d_ex_data)
clean_data <- clean4DData(raw4d_ex_data)
growthElaboration(clean_data, nweeks = 1)
}

}
