% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthProcessing.R
\name{growthProcessing}
\alias{growthProcessing}
\title{Growth data processing}
\usage{
growthProcessing(raw_data)
}
\arguments{
\item{raw_data}{The raw data downloaded with the \code{readTTData} function.}
}
\description{
This function processes the 4D data by a simplified procedure,
  to focus on the growth sensor. The data are left raw, and are calculated
  using the new equation. Week, biweek, threeweek columns are created.
  Other unuseful columns are dropped.
}
\details{
This function will be united to \code{clean4DData} when we come to
  a definitive version of the processing to be done to the growth data.
}
