% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinByDateTime.R
\name{joinByDateTime}
\alias{joinByDateTime}
\title{Join TT data with other temporal data}
\usage{
joinByDateTime(TT_data, other_data, datetime_name)
}
\arguments{
\item{TT_data}{The clean 4D Data of the tree talkers.}

\item{other_data}{The other dataframe to be joined to the TT data.}

\item{datetime_name}{The name of the column indicating date_time in the
other data.}
}
\value{
A data.frame, with the joined data.
}
\description{
This function facilitates the join of the TT data with other
  data recorded every hour. This fuction rounds the records in both
  dataframes to an hourly resolution and then joins them by date and hour by
  left join, so keeping all the rows of TT data, and leaving NA's where the
  second data.frame doesn't have records.
  The variables in the returned data.frame have the suffix ".TT" for the
  TT variables, and ".data" for the other data Joined.
}
\examples{
\dontrun{
data(raw4d_ex_data)
clean_4d <- clean4DData(raw4d_ex_data)
time_vector <- unique(clean_4d$date_hour)
external_data_example <- data.frame(
date_time_column = time_vector,
variable1 = rnorm(lenght(time_vector, 1,2))
)
colnames(external_data_example)
joined_data <- joinByDateTime(clean_4d, 
external_data_example,
datetime_name = "date_time_column")
colnames(joined_data)
print(joined_data)
}

}
