% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTreeAttributes.R
\name{joinTreeAttributes}
\alias{joinTreeAttributes}
\title{Join tree attributes}
\usage{
joinTreeAttributes(TT_data, tree_attributes)
}
\arguments{
\item{TT_data}{A data frame with the Tree Talkers 4D data, like the one returned
by \code{clean4DData}. Must include a column named id, with the TT ids}

\item{tree_attributes}{A data frame with the tree data to join. Must contain
a column id with the TT ids.
@return A joined data.frame, with suffix .TT on the columns from the TT data,
and suffix .tree on the columns from the tree attributes.}
}
\description{
Function that allows to join the TT+ data with a table of
  tree attributes. The Tree attributes table must include a column labelled
  id, with the Tree talker id associated with each tree. Can contain an
  arbitrary number of columns of other data.
}
