% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchGUI.R
\name{launchGUI}
\alias{launchGUI}
\title{Launch the GUI for the package}
\usage{
launchGUI()
}
\description{
This function launches a shiny GUI for the package, that
  allows the user to clean, explore, plot and download the TT data.
}
