% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchGUINight.R
\name{launchGUINight}
\alias{launchGUINight}
\title{Launch the GUI for the package (nighttime deltaT)}
\usage{
launchGUINight()
}
\description{
This function launches a shiny GUI for the package, that
  allows the user to clean, explore, plot and download the TT data.
}
