% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchNewGUI.R
\name{launchNewGUI}
\alias{launchNewGUI}
\title{Launch the GUI for the package}
\usage{
launchNewGUI()
}
\description{
This function launches a shiny GUI for the package, that
  allows the user to clean, explore, plot and download the TT data.
  This function is designed to deal with all 4D, 45, and 59 strings.
}
