% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTTDB.R
\name{mergeTTDB}
\alias{mergeTTDB}
\title{Merge treetalker databases}
\usage{
mergeTTDB(rawdata1, rawdata2)
}
\arguments{
\item{rawdata1}{First database.}

\item{rawdata2}{Second database.}
}
\description{
Merge two TT databases by rbinding them before processing.
  First ensures they have the same row names (and so the same length).
}
