% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainId.R
\name{obtainId}
\alias{obtainId}
\title{Obtain TT ID}
\usage{
obtainId(x)
}
\arguments{
\item{x}{The string V1 of the tree talkers data, that contains the server
date-time and the id of the Tree Talker.}
}
\value{
A character vector with the devices' ids.
}
\description{
This function retrieves the devices' ids from the first column of
  the raw Tree Talkers data. It takes as argument the first column of the TT data.
  returned by the function \code{readTTData}. The ids are returned
  as a character vector. This function works within the \code{isolateTT()}
  function, but can also be called manually.
}
\examples{
\dontrun{
data(raw4d_ex_data)
ids <- obtainId(raw4d_ex_data$V1)
print(ids)
}

}
