% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotByQuarter.R
\name{plotByQuarter}
\alias{plotByQuarter}
\title{Plot data by quarter of the year}
\usage{
plotByQuarter(clean_data, variable = "do_sap_flow", statistic = "median")
}
\arguments{
\item{clean_data}{The data frame of clean data obtained with \code{clean4DData}
or \code{h24Clean}.}

\item{variable}{The variable to be plotted}

\item{statistic}{\code{"median"} or \code{"mean"}. Defaults to median. The
statistic to be plotted.}
}
\value{
A ggplot object. Additional layers can be added to it.
}
\description{
This function plots the data by quarter of the year, similarly
  to how plotByMH works.
}
