% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDeviceFunctioning.R
\name{plotDeviceFunctioning}
\alias{plotDeviceFunctioning}
\title{Plot device functioning}
\usage{
plotDeviceFunctioning(
  TTdata4d,
  ids,
  start_date = "2021-01-01",
  end_date = "2021-12-31",
  min_valid_obs = 24
)
}
\arguments{
\item{TTdata4d}{Tree talker data processed with the clean4DData function.}

\item{ids}{A character vector of ids to be assessed. The ids must among those 
in the id column of the TTdata4d. The ids present in the data can be easily 
checked with \code{unique(TTdata4d$id)}.}

\item{start_date}{The date of beginning of the time period to be assessed.}

\item{end_date}{The date of end of the time period to be assessed.}

\item{min_valid_obs}{The number of valid observations required for each 
variable to consider the data valid. Default value = \code{24}.}
}
\value{
A ggplot object.
}
\description{
Function for plotting the assessment of the functioning of
  several devices. Takes as input the same params as 
  \code{deviceFunctioningAssesment}. Can take a character containing the ids
  to assess, and creates an ensemble plot.
}
