% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHoursTS.R
\name{plotHoursTS}
\alias{plotHoursTS}
\title{Plot single hours
 @description This function allows the user to plot the tree talker time 
   series as points. Useful while selecting single specific hours.}
\usage{
plotHoursTS(data, variable = "do_sap_flow", statistic = "median")
}
\arguments{
\item{data}{The input data (the clean table).}

\item{variable}{The variable to be represented on the y axis.}

\item{statistic}{Defaults to "median", a summary statistic to be calculated
from the data. Data are grouped by date.
Can be "median", "mean" or "none". If "none" the data of all the tree 
talkers are displayed.}
}
\description{
Plot single hours
 @description This function allows the user to plot the tree talker time 
   series as points. Useful while selecting single specific hours.
}
