% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTS.R
\name{plotTS}
\alias{plotTS}
\title{Plot the time series}
\usage{
plotTS(data, variable = "do_sap_flow", statistic = "median")
}
\arguments{
\item{data}{A data frame with the tree talkers 4D data, as the output of
the \code{read4DData} function, or furtherly modified data.}

\item{variable}{A chararacter string which specifies a valid variable
name in the data data.frame supplied to the function.}

\item{statistic}{A string with value "mean", "median" or "none". Specifies if
data shall be grouped and summarised by mean, median, or shall not be
summarised. Defaults to median, as we recommend to use the median
for the TT data, as a robust indicator.}
}
\description{
This function plots the time series of the variable selected.
  It can be run two ways: by summarising the data of all the TTs with a
  single statistic (mean or median) and plotting a single line, or by
  plotting the data of each TT as a separate coloured line.
  Also the time series of different sites can be plotted. In that case
  only a summary statistic is plotted for each site, with lines of a
  different color for each site.
}
\examples{
\dontrun{
data(raw4d_ex_data)
clean_data<- clean4DData(raw4d_ex_data,
lower.TTree = 0,
higher.TTree = 40,
lower.TAir = -5,
higher.TAir = 40)
plotTS(clean_data, variable = "Tair", statistic = "median")
filtered_data <- filterByDate(clean_data, "2021-05-10", "2021-05-20")
plotTS(filtered_data, variable = "do_sap_flow", statistic = "none")
}
}
