% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTTsDay.R
\name{plotTTsDay}
\alias{plotTTsDay}
\title{Plot number of TTs per day}
\usage{
plotTTsDay(TTdaydata, print = TRUE)
}
\arguments{
\item{TTdaydata}{The dataframe output by the \code{dailyNumberTTs} function.}

\item{print}{Default value = \code{TRUE}}
}
\value{
A ggplot object
}
\description{
Function to graphically display the output of 
  the \code{dailyNumberTTs} result. Provides a plot with the number of TTs
  available for that day with complete observations of the variable of 
  interest.
}
