% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTTData.R
\name{readTTData}
\alias{readTTData}
\title{Read TreeTalkers data from server}
\usage{
readTTData(link_string)
}
\arguments{
\item{link_string}{should be the string of the server link to the Tree Talkers data, you can use also the path to the data file if you have downloaded them}
}
\value{
an object of class "tbl" , "data.table", "data.frame"
}
\description{
Uses the fread function from data.table package with arguments set properly
}
