% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOosLight.R
\name{removeOosLight}
\alias{removeOosLight}
\title{Remove invalid light values}
\usage{
removeOosLight(x)
}
\description{
This function removes the values that are not valid from the
  dataset (the values higher than 70.000).
}
\examples{
\dontrun{
example_values <- sample(c(1:100000), size = 30)
print(example_values)
removeOosLight(example_values)
}

}
