% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot of the TT variables}
\usage{
scatterplot(
  data,
  variable1 = "Tair",
  variable2 = "do_sap_flow",
  theme = "bw",
  plot_title = "Plot",
  x_label = variable1,
  y_label = variable2
)
}
\arguments{
\item{data}{The TT+ data cleaned}

\item{variable1}{The variable to be represented on the x axis}

\item{variable2}{The variable to be represented on the y axis}

\item{theme}{The theme to be used (one of "bw", "grey", "classic" and "light")}

\item{plot_title}{The title of the plot}

\item{x_label}{The label to be put on the x axis}

\item{y_label}{The label to be put on the y axis}
}
\description{
Wrapper around ggplot, that allows the user to make a
  scatterplot of the variables recorded by the tree talkers.
}
