% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPlotEnsemble.R
\name{spectralPlotEnsemble}
\alias{spectralPlotEnsemble}
\title{TT+ spectral plot}
\usage{
spectralPlotEnsemble(
  data49,
  year = 2021,
  month = 6,
  day = 1,
  tz = Sys.timezone(),
  print = T
)
}
\arguments{
\item{data49}{A data.frame with light data from the string 49, as the one produced 
by the string49Handling function.}

\item{year}{The year of interest (numeric).}

\item{month}{The month of interest (numeric).}

\item{day}{The day of interest (numeric).}

\item{tz}{The time zone of the place where the TT+ are installed. Default
value: \code{Sys.timezone} the system timezone.}

\item{print}{a flag, set TRUE to print the plot, FALSE to only return it}
}
\description{
Function for creating a spectral plot of the TT+ data. The 
  spectral plot is a plot which has on the x axis the wavelength of the bands
  and on the Y axis the energy amount (microWatt/cm^2).
}
