% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPlotHour.R
\name{spectralPlotHour}
\alias{spectralPlotHour}
\title{Spectral hourly plot}
\usage{
spectralPlotHour(
  data49,
  year = 2021,
  month = 6,
  hour = 8,
  tz = Sys.timezone(),
  print = T
)
}
\arguments{
\item{data49}{The 49 string data. A dataframe as the one output by the
\code{string49Handling} function.}

\item{year}{Year of interest (scalar). Default: \code{2021}.}

\item{month}{Month of interest (scalar). Default: \code{6}.}

\item{hour}{Hour of interest (scalar). Default: \code{8}.}

\item{tz}{Time zone. Defaults to the System timezone.}

\item{print}{Flag \code{TRUE} or \code{FALSE}. If to print the plot or return
it.}
}
\description{
Function for plotting all the data of the same hour during a 
selected month.
}
\examples{
\dontrun{
data(raw4d_ex_data)
data49 <- string49Handling(raw4d_ex_data)
spectralPlotHour(data49, year = 2021, month = 5, hour = 9, tz = "CET")
#in the working directory a pdf file, named after the year month and hour 
#specified by the user should appear
spectralPlotHour(data49,  year = 2021, month = 5, day = 10, tz = "CET")
}
}
