% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string49Handling.R
\name{string49Handling}
\alias{string49Handling}
\title{Function for cleaning string 49 data}
\usage{
string49Handling(raw_data, time_zone = Sys.timezone())
}
\arguments{
\item{raw_data}{Raw data as those downloaded with readTTData.}

\item{time_zone}{Time zone to be used. Defaults to Sys.timezone().}
}
\description{
This function isolates the 49 string, and uniforms the data
  setting them to the same gain.
}
