% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttprocessing-package.R
\docType{package}
\name{ttprocessing-package}
\alias{ttprocessing}
\alias{ttprocessing-package}
\title{ttprocessing: Processing Tree Talker data}
\description{
Package for processing, converting and plotting the Tree Talker data.
}
\keyword{internal}
