% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttprocessing-package.R
\docType{package}
\name{ttprocessing}
\alias{ttprocessing}
\title{ttprocessing: a package for Tree Talkers data processing and plotting}
\description{
This package provides functions for downloading,
  converting, cleaning, plotting and quality check of the Tree Talkers data
}
