% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttSoilDataProcessing.R
\name{ttSoilDataProcessing}
\alias{ttSoilDataProcessing}
\title{Function for handling TT-soil data.}
\usage{
ttSoilDataProcessing(
  TTdata,
  string_type = "4D",
  equation = "S1",
  lower.TSoil = -15,
  higher.TSoil = 40,
  lower.TAir = -15,
  higher.TAir = 50,
  lower.RH = 20,
  higher.RH = 100,
  tz = Sys.timezone()
)
}
\arguments{
\item{TTdata}{Raw TT+ data, containing the TT-soil 45, 4D or 55 string.}

\item{string_type}{The type of string produced by the TT-soil. Can be:
\code{"4D"}, \code{"55"} or \code{"59"}. Default value: "4D".}

\item{equation}{The equation from the manual to be used. Possible
names of equations are: \code{"S1"}, \code{"S2"}, \code{"S3"},
\code{"S4"}, \code{"S5"}, \code{"S6"} or \code{"S7"}.}

\item{lower.TSoil}{Lower threshold for the soil temperature records. Values
,lower than this will be discarded.}

\item{higher.TSoil}{Higher threshold for the soil temperature records. Values
higher than this will be discarded.}

\item{lower.TAir}{Lower threshold for the air temperature records. Values
lower than this will be discarded.}

\item{higher.TAir}{Higher threshold for the air temperature records. Values
higher than this will be discarded.}

\item{lower.RH}{Lower threshold for the relative humidity. Values lower than
this will be discarded.}

\item{higher.RH}{Higher threshold for the relative humidity. Values higher
than this will be discarded.}

\item{tz}{Timezone. Defaults to system timezone.}
}
\description{
This function implements the calibration equations from the
TT-soil manual, and is able to handle the TT-soil data according to them.
}
\details{
For handling the TT-soil data, the user has to choose the equation
that matches the probe type and the soil bulk density at the study site.
Than the raw data are processed the following way:
\enumerate{
\item Temperature is converted from DN to degrees C with the lookup table.
\item The values out of range of the raw Ecf are replaced with \code{NA}
\item It is corrected for temperature sensitivity
\item The Ecf corrected is used for estimating soil water content.
}

Table with equation coefficients that are used in the equations.\tabular{lccccccc}{
   Equation \tab Probe freq. \tab Soil bulk density \tab m (T sensitivity) \tab b (T sensitivity) \tab B1 \tab B2 \tab B3 \cr
   S1 \tab 50000 \tab 1.622 \tab -5.6 \tab 830 \tab 3.08x10^-7 \tab -0.005394 \tab 26 \cr
   S2 \tab 50000 \tab 1.194 \tab -5.6 \tab 830 \tab 1.30x10^-8 \tab -0.0009713 \tab 31 \cr
   S3 \tab 3000000 \tab 0.669 \tab -758 \tab 154844 \tab 1.9x10^-11 \tab -6.16x10^-5 \tab 59 \cr
   S4 \tab 3000000 \tab 1.221 \tab -758 \tab 154844 \tab 6.4x10^-12 \tab -1.93x10^-5 \tab 49 \cr
   S5 \tab 3000000 \tab 0.9708 \tab -758 \tab 154844 \tab 7.1x10^-12 \tab -1.78x10^-5 \tab 17 \cr
   S6 \tab 3000000 \tab 0.871 \tab -758 \tab 154844 \tab 5.7x10^-10 \tab -0.0003022 \tab 53 \cr
   S7 \tab 3000000 \tab 0.3866 \tab -758 \tab 154844 \tab 1.68x10^-9 \tab -0.000521 \tab 41 \cr
}
}
\examples{
\dontrun{
raw_data <- readTTdata("path or link to the data")
#run the function on the raw data, and select the proper equation,
#based on the bulk density of the soil, and on the frequency of the probe
#supported by your device
ttsoil_processed_data  <- ttSoilDataProcessing(raw_data,
string_type = "4D", equation = "S1") \%>\%
 filterIDs("id of the TT-soil")
print(ttsoil_processed_data)
colnames(ttsoil_processed_data)
}


}
