% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzConvert.R
\name{tzConvert}
\alias{tzConvert}
\title{tzConvert converts the time to local time}
\usage{
tzConvert(t, tz = Sys.timezone())
}
\arguments{
\item{t}{A time vector containing the POSIXct time of the observations.}

\item{tz}{Defaults to the timezone set on the device used.
Can be set manually to any timezone accepted by lubridate}
}
\value{
A time vector "POSIXct" with the local time corresponding to the observations
}
\description{
This function converts the time from UTC as recorded by the Tree Talkers to the user's local time. It is used mainly as a helper function in clean4DData(), but can also be used for cleaning the data manually.
}
\examples{
\dontrun{
example_time <- Sys.time()
tzConvert(example_time, tz = "UTC")
}

}
