% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPD.R
\name{VPD}
\alias{VPD}
\title{Vapour Pressure Deficit}
\usage{
VPD(Tair, RH)
}
\arguments{
\item{Tair}{A numeric vector containing the Tair values}

\item{RH}{A numeric vector containing relative humidity data. Must be the same length
as Tair.}
}
\description{
This function calculates Vapour Pressure Deficit using the air
  temperature and relative humidity data recorded by the Tree Talker. It takes
  as input two numeric vectors.
}
\examples{
\dontrun{
#compute the VPD from temperature and air humidity
VPD(Tair = 10, RH = 50)
VPD(Tair = 20, RH = 80)
}


}
