% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTemperatureTree.R
\name{cleanTemperatureTree}
\alias{cleanTemperatureTree}
\title{Clean the temperature recorded inside the trunk}
\usage{
cleanTemperatureTree(x, lower.TTree = -10, higher.TTree = 40)
}
\arguments{
\item{x}{Temperature in degrees C, as obtained by \code{convertTemperature}. A
numeric vector.}

\item{lower.TTree}{A value for the lowest temperature value accepted.}

\item{higher.TTree}{A value for the highest temperature value accepted.}
}
\value{
A numeric vector, with values outside the range replaced with NA's.
}
\description{
For the purpose of measuring the sap flow, temperature is measured
  by two probes put inside the trunk. Those values are used later for calculating
  the sap flow. For a correct calculation it is important to remove values
  that are not plausible. We have chosen to remove temperatures under 0 degrees 
  C and
  above 40 degrees C, but the function allows the user to choose manually the range
  of temperature he regards as acceptable. Values outside the range are replaced
  with NA.
}
