% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTemperature.R
\name{convertTemperature}
\alias{convertTemperature}
\title{Convert trunk temperature from DN to °C}
\usage{
convertTemperature(x)
}
\arguments{
\item{x}{A numeric vector containing the DN of the trunk temperature measurment}
}
\description{
Converts the temperature measured inside the trunk from DN to °C
 following the equation in the TT+ manual. Works mainly as a helper function
 inside clean4DData, but can also be called manually.
}
