% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyQualityCheck.R
\name{dailyQualityCheck}
\alias{dailyQualityCheck}
\title{Data quality check}
\usage{
dailyQualityCheck(TTdata4d)
}
\arguments{
\item{TTdata4d}{Dataframe with Tree talker 4D data, such as the one produced
by \code{clean4DData}.}
}
\description{
Checks the data quality by counting the number of daily
  valid records each day for each tree talker for each variable.
  Returns a data.frame with these data.
}
