% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastBatteryCheck.R
\name{fastBatteryCheck}
\alias{fastBatteryCheck}
\title{Check battery level}
\usage{
fastBatteryCheck(link_string, ndays_ago = 60)
}
\arguments{
\item{link_string}{Link to the cloud for which the battery levels are to be
checked.}

\item{ndays_ago}{Number of days ago to consider. Defaults to \code{60}.}
}
\description{
function to fastly check the battery level of the devices.
}
