% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByHour.R
\name{filterByHour}
\alias{filterByHour}
\title{Filter by hour}
\usage{
filterByHour(data4D, hours)
}
\arguments{
\item{data4D}{The 4D TT+ data}

\item{hours}{A character vectors of the hours to be included: e.g.:
c("2","3","4") or "12".}
}
\description{
filter that allows the user to filter the TT data by hour.
}
\examples{
\dontrun{
data(raw4d_ex_data)
clean_data <- clean4DData(raw4d_ex_data)
plotByMH(clean_data)
filtered_data <-  filterByHour(clean_data, c("8", "9", "10", "11", "12", "13", "14", "15", "16"))
plotByMH(filtered_data)
}
}
