% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join4D49.R
\name{join4D49}
\alias{join4D49}
\title{Join string 4D with string 49}
\usage{
join4D49(data4D, data49)
}
\arguments{
\item{data4D}{The 4d data data.frame.}

\item{data49}{The 49 data data.frame.}
}
\description{
Joins the string 4D and the string 49 by left join, using the
  \code{dplyr} \code{left_join} function. Joins by id and date_hour of the
  observation.
}
