% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchGUINight.R
\name{launchGUINight}
\alias{launchGUINight}
\title{Launch the GUI for the package (nighttime deltaT)}
\usage{
launchGUINight()
}
\description{
This function launches a shiny GUI for the package, that
  allows the user to clean, explore, plot and download the TT data.
}
\examples{
\dontrun{
 # example code
 # The GUI can read TT+ data from a remote or local txt file. 
 #the txt file has to be formatted the same way as the TT+ data in the server.
 # to try it, the txt files with TT+ can be used from the github repository of 
 # this package.
 # If not, the example data can be saved to a local csv file 
 library(readr)
 data(raw4d_ex_data)
 #create a local file with data
 write_csv2(raw4d_ex_data, file = "filename.txt", col_names = F)
 # after launching the GUI, the path to the file created with write_csv2
 #has to be put as text input in the first panel, and then the data loading
 #button has to be pressed
 #after the data appear, they can be processed in the data cleaning tab
 #they can be saved to CSV, or plotted within the plotting tabs.
 launchGUINight()
 

}

}
