% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSites.R
\name{mergeSites}
\alias{mergeSites}
\title{Merge sites}
\usage{
mergeSites(
  TT_data_frames,
  site_names = as.character(c(1:length(TT_data_frames)))
)
}
\description{
This function allows the user to merge the clean TT data
  of some sites together. The function is a wrapper around the dplyr
   \code{bind_rows} function.
}
