% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSites.R
\name{mergeSites}
\alias{mergeSites}
\title{Merge sites}
\usage{
mergeSites(
  TT_data_frames,
  site_names = as.character(c(1:length(TT_data_frames)))
)
}
\arguments{
\item{TT_data_frames}{list of TT data frames to be merged.}

\item{site_names}{Names of the sites to which the data.frames belong. 
If not provided, by default, they will be named with progressive numbers in 
the order they were supplied.}
}
\description{
This function allows the user to merge the clean TT data
  of some sites together. The function is a wrapper around the dplyr
   \code{bind_rows} function.
}
\examples{

\dontrun{

data("raw4d_ex_data")
#simulate a database from a second site by duplicating the example data
example_site_2 <- raw4d_ex_data
clean_1 <- clean4DData(raw4d_ex_data)
clean_2 <- clean4DData(example_site_2)
nrow(clean_1)
nrow(clean_2)
list_of_databases <- list(clean_1, clean_2)
sitenames <- c("site1", "site2")
merged_database <- mergeSites(list_of_databases, sitenames)
nrow(merged_database)
unique(merged_database$site)
print(merged_database)
summary(merged_database)

}

  
}
