% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTTDB.R
\name{mergeTTDB}
\alias{mergeTTDB}
\title{Merge treetalker databases}
\usage{
mergeTTDB(rawdata1, rawdata2)
}
\arguments{
\item{rawdata1}{First database.}

\item{rawdata2}{Second database.}
}
\description{
Merge two TT databases by rbinding them before processing.
  First ensures they have the same row names (and so the same length).
}
\examples{
\dontrun{
data(raw4d_ex_data)
#simulate two databases by splitting the example data in 2 part
example_db1 <- raw4d_ex_data[V4 < 1623837601,]
example_db2 <- raw4d_ex_data[V4 >= 1623837601,]
nrow(example_db1)
nrow(example_db2)
summary(example_db1)
summary(example_db2)
merged_database <- mergeTTDB(example_db1, example_db2)
print(merged_database)
summary(merged_database)
}

}
