% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotByMH.R
\name{plotByMH}
\alias{plotByMH}
\title{Plot 4D data by month and hour}
\usage{
plotByMH(clean_data, variable = "do_sap_flow", statistic = "median")
}
\arguments{
\item{clean_data}{A data frame with TT data, as those created by the
functions of this package as for example \code{read4DData}.}

\item{variable}{The variable to plot. It must be one of the names of the
input data.frame.}

\item{statistic}{The statistic to be plotted. Can be one of "mean" and
"median". We have chosen "median" as default, as the TT data aren't always
very clean, so the median will be a more robust indicator, and we suggest
to use it.}
}
\value{
a ggplot object, that can be further customized.
}
\description{
This function is a wrapper around ggplot, to display
  the TT data by month and hour, helping in EDA. The function returns
  a ggplot object, to which can be added other layers such as those to
  specify the theme. The plot is also printed.
}
\details{
The summary statistic is calculated with na.rm = T, as it is very
  likely that the data will contain NA.
}
