% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrowth.R
\name{plotGrowth}
\alias{plotGrowth}
\title{Plot elaborated growth data}
\usage{
plotGrowth(growth_data, variable = "growth_s0", theme = "bw")
}
\arguments{
\item{growth_data}{The growth data obtained with the \code{growthElaboration}
function.}

\item{variable}{a string with the name of the variable to be plotted among
the column names in the growth_data object. One of: "growth_mode_r",
"growth_s0", "growth_mode_cm".}

\item{theme}{The theme to be used for the plot. One of: "bw", "classic",
"grey" or "light" to select the corresponding ggplot theme.}
}
\description{
Function for plotting the growth data obtained with the
  \code{growthElaboration} function.
}
