% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHoursTS.R
\name{plotHoursTS}
\alias{plotHoursTS}
\title{Plot single hours
 @description This function allows the user to plot the tree talker time 
   series as points. Useful while selecting single specific hours.}
\usage{
plotHoursTS(data, variable = "do_sap_flow", statistic = "median")
}
\arguments{
\item{data}{The input data (the clean table).}

\item{variable}{The variable to be represented on the y axis.}

\item{statistic}{Defaults to "median", a summary statistic to be calculated
from the data. Data are grouped by date.
Can be "median", "mean" or "none". If "none" the data of all the tree 
talkers are displayed.}
}
\description{
Plot single hours
 @description This function allows the user to plot the tree talker time 
   series as points. Useful while selecting single specific hours.
}
\examples{
\dontrun{
data(raw4d_ex_data)
clean_data<- clean4DData(raw4d_ex_data,
lower.TTree = 0,
higher.TTree = 40,
lower.TAir = -5,
higher.TAir = 40)
filtered_data <- filterByHour(clean_data, hours = c("11", "12", "13"))
plotHoursTS(filtered_data, variable = "do_sap_flow", statistic = "none")
filtered_data2 <- filterByHour(clean_data, hours = c("13"))
plotHoursTS(filtered_data, variable = "Tair", statistic = "median")
}
}
