% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWithShade.R
\name{plotWithShade}
\alias{plotWithShade}
\title{Plot shaded Timeseries}
\usage{
plotWithShade(
  data,
  variable = "do_sap_flow",
  statistic = "median",
  lat = 40,
  lon = 14
)
}
\arguments{
\item{data}{The clean 4D data of the tree talkers}

\item{variable}{The variable to be plotted. Same as plot TS (this is a
a variant of the plotTS function)}

\item{statistic}{The summary arg. Must be one of "mean" or "median".
The default is "median".}
}
\description{
Plots the timeseries of TT+ data with nights shaded, by using
  the \code{plotTS} function, adding some ggplot layers and retrieving
  sunrise and sunset from the suncalc library.
}
